[package] Add DependencyExplorer
**Summary**
This commit adds `DependencyExplorer`, a class that can be used to
explore and understand the dependencies of a package prior to packaging
it using `PackageExporter`.

The dependency analysis for a given module/object is obtained from a
`PackageExporter` instance. This class has been modified to accept a
`raise_packaging_errors` argument that, if `True`, does not throw an
exception when either a module or its source are not found. Instead, it
records these "broken" modules and the reason they are broken in an
attribute. A `DependencyExplorer` instance can then extract dependency
information from a `PackageExporter` instance it owns by accessing
`debug_deps` (the dependency graph) and `broken_modules` (the modules
that need to be eliminated from the dependency graph via externing or
mocking).

`DependencyResolver` can check if a module or object is packageable
(`can_package`), can return a list of dependencies that need to be
addressed somehow using `mock`/`extern` (`get_unresolved_dependencies`),
and can visualize the dependencies (`show_dependency_graph`). More
functionality (e.g. getting suggestions on what to `extern`/`mock` using
some fancy graph analysis, graph pruning via "third party" module
declarations) can be added on this foundation in the near future.

**Test Plan**
This commit adds a unit test (`test_dependency_explorer`) that tests
`can_package` and `get_unresolved_dependencies`. I used the
`DependencyResolver` instance in an interactive terminal session to test
`show_dependency_graph`.

ghstack-source-id: ec8552a048dcfc97bb2f45b13b773600a46ad534
Pull Request resolved: https://github.com/pytorch/pytorch/pull/56234